+++ src/interface/interactive.py
@@ -23,9 +23,6 @@ from lib.Item import Item
 
 import hashlib
 
-"""This example implements the interaction between Qt Widgets and a 3D
-matplotlib plot"""
-
 
 def init(data) -> list:
     items = []
@@ -42,7 +39,7 @@ class ApplicationWindow(QMainWindow):
         logi = [0, 0, 0]
         self.container = Container(0, 0, 0, logi, 100, 10, 5)
 
-        self.column_names = ["Column A", "Column B", "Column C"]
+        self.column_names = ["Color", "Serial"]
 
         # Central widget
         self._main = QWidget()
@@ -85,10 +82,9 @@ class ApplicationWindow(QMainWindow):
         # ComboBox (Right)
         self.combo = QComboBox()
         self.load_list()
-        # items = ["Wired", "Surface", "Triangular Surface", "Sphere", "Cube"]
         items = []
         for i in range(len(self.shipment)):
-            items.append(f"Crate {1+i}")
+            items.append(f"Crate {1 + i}")
         self.combo.addItems(items)
 
         # Right layout
@@ -154,25 +150,13 @@ class ApplicationWindow(QMainWindow):
             color = QColor()
             hex_code = f'#{self.serial_hash(data[i].item.get_serial())}'
             color.setNamedColor(hex_code)
-            print(hex_code)
-            self.table.setItem(table_row, 0, QTableWidgetItem())
-            self.table.item(table_row, 0).setBackground(color)
+            print(hex_code, table_row)
+            item = QTableWidgetItem("new Item").setBackground(color)
+            self.table.setItem(table_row, 0, item)
+            # self.table.item(table_row, 0)
             self.table.setItem(table_row, 1, QTableWidgetItem(data[i].item.get_serial()))
             table_row += 1
             seen.append(data[i].item.get_serial())
-
-    def set_canvas_table_configuration(self, row_count, data):
-        self.fig.set_canvas(self.canvas)
-        self._ax = self.canvas.figure.add_subplot(projection="3d")
-
-        self._ax.set_xlabel(self.column_names[0])
-        self._ax.set_ylabel(self.column_names[1])
-        self._ax.set_zlabel(self.column_names[2])
-
-        self.table.setRowCount(row_count)
-        self.table.setColumnCount(3)
-        self.table.setHorizontalHeaderLabels(self.column_names)
-        self.set_table_data(data[0], data[1], data[2])
 
     def set_canvas_table_configuration_containers(self, row_count, data):
         self.fig.set_canvas(self.canvas)
@@ -188,69 +172,18 @@ class ApplicationWindow(QMainWindow):
         sha.update(item_serial.encode())
         return sha.hexdigest()[:6]
 
-    def load_list(self):
-        self.data = Ingest.ingest("../", 'Book1.xlsx')
+    def load_list(self, packing_list="Book1.xlsx"):
+        self.data = Ingest.ingest("../", packing_list)
         self.items = init(self.data)
         self.items = Sort.item_sort(self.items)
         print(self.items[0])
         print(self.items[len(self.items) - 1])
         logi = [0, 0, 0]
-        self.containerTemplate = Container(0, 0, 0, logi, 1087, 1277, 980)
+        self.containerTemplate = Container(0, 0, 0, logi, 2000, 2000, 2500)
         self.shipment = palletize.palletize(self.items, self.containerTemplate)
         print(self.shipment)
 
     # Plot methods
-
-    def plot_wire(self):
-        # Data
-        self.X, self.Y, self.Z = axes3d.get_test_data(0.03)
-
-        self.set_canvas_table_configuration(len(self.X[0]), (self.X[0], self.Y[0], self.Z[0]))
-        self._ax.plot_wireframe(self.X, self.Y, self.Z, rstride=10, cstride=10, cmap="viridis")
-        self.canvas.draw()
-
-    def plot_surface(self):
-        # Data
-        self.X, self.Y = np.meshgrid(np.linspace(-6, 6, 30), np.linspace(-6, 6, 30))
-        self.Z = np.sin(np.sqrt(self.X ** 2 + self.Y ** 2))
-
-        self.set_canvas_table_configuration(len(self.X[0]), (self.X[0], self.Y[0], self.Z[0]))
-        self._ax.plot_surface(self.X, self.Y, self.Z,
-                              rstride=1, cstride=1, cmap="viridis", edgecolor="none")
-        self.canvas.draw()
-
-    def plot_triangular_surface(self):
-        # Data
-        radii = np.linspace(0.125, 1.0, 8)
-        angles = np.linspace(0, 2 * np.pi, 36, endpoint=False)[..., np.newaxis]
-        self.X = np.append(0, (radii * np.cos(angles)).flatten())
-        self.Y = np.append(0, (radii * np.sin(angles)).flatten())
-        self.Z = np.sin(-self.X * self.Y)
-
-        self.set_canvas_table_configuration(len(self.X), (self.X, self.Y, self.Z))
-        self._ax.plot_trisurf(self.X, self.Y, self.Z, linewidth=0.2, antialiased=True)
-        self.canvas.draw()
-
-    def plot_sphere(self):
-        # Data
-        u = np.linspace(0, 2 * np.pi, 100)
-        v = np.linspace(0, np.pi, 100)
-        self.X = 10 * np.outer(np.cos(u), np.sin(v))
-        self.Y = 10 * np.outer(np.sin(u), np.sin(v))
-        self.Z = 9 * np.outer(np.ones(np.size(u)), np.cos(v))
-
-        self.set_canvas_table_configuration(len(self.X), (self.X[0], self.Y[0], self.Z[0]))
-        self._ax.plot_surface(self.X, self.Y, self.Z)
-        self.canvas.draw()
-
-    def plot_cube(self):
-        self._ax.cla()
-        r = [-10, 10]
-        for s, e in combinations(np.array(list(product(r, r, r))), 2):
-            if np.sum(np.abs(s - e)) == r[1] - r[0]:
-                self._ax.plot3D(*zip(s, e), color="b")
-
-        self.canvas.draw()
 
     def plot_dyn_cube(self, shipment: list, index: int):
         if hasattr(self, "_ax"):
@@ -262,14 +195,16 @@ class ApplicationWindow(QMainWindow):
             if container.item is None:
                 continue
             verts = np.array([
-                [container.x, container.y, container.z],                                                                                         # 0
-                [container.x+container.item.get_length(), container.y, container.z],                                                             # 1
-                [container.x, container.y+container.item.get_width(), container.z],                                                              # 2
-                [container.x+container.item.get_length(), container.y+container.item.get_width(), container.z],                                  # 3
-                [container.x, container.y, container.z+container.item.get_height()],                                                             # 4
-                [container.x + container.item.get_length(), container.y, container.z+container.item.get_height()],                               # 5
-                [container.x, container.y + container.item.get_width(), container.z+container.item.get_height()],                                # 6
-                [container.x + container.item.get_length(), container.y + container.item.get_width(), container.z+container.item.get_height()],  # 7
+                [container.x, container.y, container.z],  # 0
+                [container.x + container.item.get_length(), container.y, container.z],  # 1
+                [container.x, container.y + container.item.get_width(), container.z],  # 2
+                [container.x + container.item.get_length(), container.y + container.item.get_width(), container.z],  # 3
+                [container.x, container.y, container.z + container.item.get_height()],  # 4
+                [container.x + container.item.get_length(), container.y, container.z + container.item.get_height()],
+                # 5
+                [container.x, container.y + container.item.get_width(), container.z + container.item.get_height()],  # 6
+                [container.x + container.item.get_length(), container.y + container.item.get_width(),
+                 container.z + container.item.get_height()],  # 7
             ])
             faces = [
                 [verts[0], verts[1], verts[3], verts[2]],  # Top
@@ -277,22 +212,31 @@ class ApplicationWindow(QMainWindow):
                 [verts[2], verts[3], verts[7], verts[6]],  # North
                 [verts[0], verts[1], verts[5], verts[4]],  # South
                 [verts[0], verts[2], verts[6], verts[4]],  # East
-                [verts[1], verts[3], verts[7], verts[5]]   # West
+                [verts[1], verts[3], verts[7], verts[5]]  # West
             ]
 
             self._ax.scatter3D(verts[:, 0], verts[:, 1], verts[:, 2], alpha=0)
             color = self.serial_hash(container.item.get_serial())
-            self._ax.add_collection3d(Poly3DCollection(faces, facecolors=f'#{color}', linewidths=1, edgecolors='b', alpha=.5))
+            self._ax.add_collection3d(
+                Poly3DCollection(faces, facecolors=f'#{color}', linewidths=1, edgecolors='b', alpha=.5))
 
         verts = np.array([
             [self.containerTemplate.x, self.containerTemplate.y, self.containerTemplate.z],  # 0
-            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y, self.containerTemplate.z],  # 1
-            [self.containerTemplate.x, self.containerTemplate.y + self.containerTemplate.width, self.containerTemplate.z],  # 2
-            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y + self.containerTemplate.width, self.containerTemplate.z],  # 3
-            [self.containerTemplate.x, self.containerTemplate.y, self.containerTemplate.z + self.containerTemplate.height],  # 4
-            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y, self.containerTemplate.z + self.containerTemplate.height],  # 5
-            [self.containerTemplate.x, self.containerTemplate.y + self.containerTemplate.width, self.containerTemplate.z + self.containerTemplate.height],  # 6
-            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y + self.containerTemplate.width, self.containerTemplate.z + self.containerTemplate.height],  # 7
+            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y,
+             self.containerTemplate.z],  # 1
+            [self.containerTemplate.x, self.containerTemplate.y + self.containerTemplate.width,
+             self.containerTemplate.z],  # 2
+            [self.containerTemplate.x + self.containerTemplate.length,
+             self.containerTemplate.y + self.containerTemplate.width, self.containerTemplate.z],  # 3
+            [self.containerTemplate.x, self.containerTemplate.y,
+             self.containerTemplate.z + self.containerTemplate.height],  # 4
+            [self.containerTemplate.x + self.containerTemplate.length, self.containerTemplate.y,
+             self.containerTemplate.z + self.containerTemplate.height],  # 5
+            [self.containerTemplate.x, self.containerTemplate.y + self.containerTemplate.width,
+             self.containerTemplate.z + self.containerTemplate.height],  # 6
+            [self.containerTemplate.x + self.containerTemplate.length,
+             self.containerTemplate.y + self.containerTemplate.width,
+             self.containerTemplate.z + self.containerTemplate.height],  # 7
         ])
         self._ax.scatter3D(verts[:, 0], verts[:, 1], verts[:, 2])
 
@@ -302,19 +246,8 @@ class ApplicationWindow(QMainWindow):
         self.canvas.draw()
 
     # Slots
-
     @Slot()
     def combo_option(self, text):
-        if text == "Wired":
-            self.plot_wire()
-        elif text == "Surface":
-            self.plot_surface()
-        elif text == "Triangular Surface":
-            self.plot_triangular_surface()
-        elif text == "Sphere":
-            self.plot_sphere()
-        elif text == "Cube":
-            self.plot_cube()
         for i in range(len(self.shipment)):
             if text == f"Crate {1 + i}":
                 self.plot_dyn_cube(self.shipment, i)