+++ changelog.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2.2] - 4-18-21
+### Added
+- analytics file to compute the packing efficiency.
+### Changed
+- Removed several lines that were not being used
+- Deprecated some methods of container class.
 
 ## [0.2.1] - 4-15-21
 ### Added